+++ k8s-configs/cluster-tools/base/pgo/base/update-pgo.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 set -e
 
+# NOTE: PGO does not version postgres-operator-examples, so most likely it is the latest sha that you will pass into
+# this script... their versioning is quite strange and does not show up in the main PGO repo either.
+# TODO: improve the versioning so it is less hacky (requires some research into how PGO is versioning behind the scenes)
+
 if [ ! $# -eq 1 ];then
+   echo "Usage: commit_sha is the corresponding commit you would like to upgrade to in the postgres-operator-examples repo"
    echo "pass arguments: to $0 <commit_sha>"
    exit 1
 fi
@@ -11,13 +16,11 @@ if [[ ! "$(pwd)" =~ "k8s-configs/cluster-tools/base/pgo/base" ]]; then
     exit 1
 fi
 
-# TODO: take a SHA to pull a specific version of the example repo
-
 source ../../../../../utils.sh
 
 # Update the PGO CRDs, other resources based on the github.com/CrunchyData/postgres-operator-examples repo.
 # NOTE: only run this script in the k8s-configs/cluster-tools/base/pgo/base directory
-cur_date=$(date -I seconds)
+cur_date=$(date -Iseconds)
 tmp_dir="/tmp/pgo/${cur_date}"
 example_repo="postgres-operator-examples"
 commit_sha="${1}"
@@ -28,8 +31,9 @@ log "Creating tmp dir - ${tmp_dir}"
 mkdir -p "${tmp_dir}"
 
 git clone "https://github.com/CrunchyData/${example_repo}" "${tmp_dir}/${example_repo}"
-cd ${tmp_dir}/${example_repo}
+pushd ${tmp_dir}/${example_repo}
 git reset --hard ${commit_sha}
+popd
 
 # Remove the singlenamespace dir, we don't use it
 rm -rf "${repo_dir_kustomize}/singlenamespace"