+++ tests/python-utils/k8s_utils.py
@@ -142,21 +142,74 @@ class K8sUtils(unittest.TestCase):
 
     def get_namespace_names(self):
         namespaces = self.core_client.list_namespace()
-        return [
-            ns.metadata.name
-            for ns in namespaces.items
-        ]
+        return [ns.metadata.name for ns in namespaces.items]
 
-    def get_namespaced_pod_names(self, namespace: str, pod_name_pattern: str) -> [str]:
+    def get_first_matching_pod_name(self, namespace: str, label: str) -> str:
+        try:
+            return next(
+                name
+                for name in self.get_deployment_pod_names(label, namespace)
+            )
+        except StopIteration:
+            print(
+                f"Pod not found for label {label} in namespace {namespace}"
+            )
+            return ""
+
+    def get_pod_env_vars(self, namespace: str, label: str) -> [str]:
         """
-        Get a list of pod_names for pods in a namespace that match a naming pattern
-        :param namespace: Namespace to check pod names
-        :param pod_name_pattern: Regex pod name pattern to check against pod names
-        :returns: {pod_name: pod_IP}
+        Exec into a pod and get the environment variables
+
+        :param namespace: Namespace of pod
+        :param label: Pod label
+        :return: List of environment variables
         """
-        pods = self.core_client.list_namespaced_pod(namespace)
-        return [
-            pod.metadata.name
-            for pod in pods.items
-            if re.search(pod_name_pattern, pod.metadata.name)
-        ]
+        pod_name = self.get_first_matching_pod_name(namespace, label)
+        if not pod_name:
+            return []
+        return self.exec_command(namespace, pod_name, ["env"]).split("\n")
+
+    def run_python_script_in_pod(
+        self, namespace: str, label: str, script_path: str
+    ) -> [str]:
+        pod_name = self.get_first_matching_pod_name(namespace, label)
+        if not pod_name:
+            return []
+        return self.exec_command(namespace, pod_name, ["python", script_path]).split(
+            "\n"
+        )
+
+    def exec_command(self, namespace: str, pod_name: str, command: [str]) -> str:
+        """
+        Execute a command in a running pod
+
+        :param namespace: Namespace of pod
+        :param pod_name: Pod name
+        :param command: Command to run in pod
+        :return: response
+        """
+        try:
+            res = k8s.stream.stream(
+                self.core_client.connect_get_namespaced_pod_exec,
+                pod_name,
+                namespace,
+                command=command,
+                stderr=True,
+                stdin=False,
+                stdout=True,
+                tty=False,
+            )
+            return res
+        except k8s.client.exceptions.ApiException as err:
+            print(f"Unable to exec in pod {pod_name}. {err}")
+            return ""
+
+    def get_configmap_values(self, namespace: str, configmap_name: str) -> {str: str}:
+        try:
+            res = self.core_client.read_namespaced_config_map(configmap_name, namespace)
+            return res.data
+        except k8s.client.exceptions.ApiException as err:
+            print(
+                f"Unable to get values for configmap {configmap_name} in namespace {namespace}. {err}"
+            )
+            return {}